use better type

+++ osmomath/decimal.go
@@ -47,7 +47,7 @@ var (
 	// precisionFactorSDK is used to adjust the scale of BigDec values to match the precision expected by sdk.Dec
 	precisionFactorSDK *big.Int
 	// precisionDiffFromSDKDec is a difference in precision between BigDec and sdk.Dec
-	precisionDiffFromSDKDec int
+	precisionDiffFromSDKDec int64
 
 	// log_2(e)
 	// From: https://www.wolframalpha.com/input?i=log_2%28e%29+with+37+digits
@@ -80,7 +80,7 @@ func init() {
 		panic("invalid decimal precision")
 	}
 
-	precisionFactorSDK = new(big.Int).Exp(big.NewInt(10), big.NewInt(int64(precisionDiffFromSDKDec)), nil)
+	precisionFactorSDK = new(big.Int).Exp(big.NewInt(10), big.NewInt(precisionDiffFromSDKDec), nil)
 }
 
 func precisionInt() *big.Int {
